1. Optimize mapreduce iteration performance 2. Create map reduce folder

+++ HW2_102062111_MR/build_map_A/execute.sh
@@ -1,5 +1,10 @@
-output=HW2/map_input
-input=/shared/HW2/sample-in/input-100M
+if [ $# -lt 1 ]; then 
+  echo "please input size : ./execute.sh \S{SIZE}"
+  exit 1
+fi
+
+output=HW2_$1/map_input
+input=/shared/HW2/sample-in/input-$1
 
 hdfs dfs -rm -r ${output}
 hadoop jar CalculateAverage.jar calculateAverage.CalculateAverage $input $output
