+++ HW2_102062111_MR/build_map_B/execute.sh
@@ -1,5 +1,10 @@
-output=HW2/iterator0C
-input=HW2/map_input
+if [ $# -lt 1 ]; then 
+  echo "please input size : ./execute.sh \S{SIZE}"
+  exit 1
+fi
+
+output=HW2_$1/iterator0C
+input=HW2_$1/map_input
 
 hdfs dfs -rm -r -f $output
 hadoop jar CalculateAverage.jar calculateAverage.CalculateAverage $input $output
